Remove question and add make choice tag for multipleSelect
